added fractional numbers to input

+++ src/main/java/com/dimonandpumba/calculator/Calculator.java
@@ -2,23 +2,23 @@ package main.java.com.dimonandpumba.calculator;
 
 public class Calculator implements ICalculator {
     @Override
-    public double add(int x, int y) {
+    public double add(double x, double y) {
         return x+y;
     }
 
     @Override
-    public double minus(int x, int y) {
+    public double minus(double x, double y) {
         return x-y;
     }
 
     @Override
-    public double multiply(int x, int y) {
+    public double multiply(double x,double y) {
         return x*y;
     }
 
     @Override
-    public double divide(int x, int y) {
+    public double divide(double x, double y) {
         if(y==0) throw new ArithmeticException();
-        return (double)x/y;
+        return x/y;
     }
 }
